refactor(config): remove AWS vocabulary from generic config schemas
Move AMI resolution cache config into the AWS provider package and drop the
AWS-specific resource-prefix fields (launch_template, instance, fleet,
spot_fleet, asg) from the shared ResourcePrefixConfig/PrefixConfig. Prefix
schemas now allow extra keys so provider-supplied prefixes still resolve via
the hasattr-based get_resource_prefix dispatch, falling back to default_prefix
when unset. Generic naming patterns no longer enumerate provider identifiers;
each provider contributes its own. This keeps the core config layer free of
provider knowledge.
