+++ src/orb/providers/aws/configuration/__init__.py
@@ -1,6 +1,7 @@
 """AWS provider configuration."""
 
 from .batch_sizes_config import AWSBatchSizesConfig
+from .caching_config import AMIResolutionCacheConfig
 from .cleanup_config import CleanupConfig, CleanupResourcesConfig
 from .config import (
     AWSProviderConfig,
@@ -12,6 +13,7 @@ from .naming_config import AWSNamingConfig
 from .validator import AWSConfigManager, get_aws_config_manager
 
 __all__: list[str] = [
+    "AMIResolutionCacheConfig",
     "AWSBatchSizesConfig",
     "AWSConfigManager",
     "AWSNamingConfig",
